move schema check openapi tests into feature directories (#35932)
